// packages/generator/src/emitter/dmmf/types.ts
export type ReadonlyDeep<O> = {
  +readonly [K in keyof O]: ReadonlyDeep<O[K]>;
};

export namespace DMMF {
// ...
  export type SchemaArgInputType = ReadonlyDeep<{
    isList: boolean;
    // type: ArgType;
    location: FieldLocation;
    namespace?: FieldNamespace;
    // additional props
    type: string;
  }>;
  export type SchemaArg = ReadonlyDeep<{
    name: string;
    comment?: string;
    isNullable: boolean;
    isRequired: boolean;
    // inputTypes: InputTypeRef[];
    deprecation?: Deprecation;
    // additional props:
    selectedInputType: SchemaArgInputType;
    typeName: string;
    typeGraphQLType: string;
    fieldTSType: string;
    hasMappedName: boolean;
    isOmitted: boolean;
  }>;
// ...
  export type Deprecation = ReadonlyDeep<{
    sinceVersion: string;
    reason: string;
    plannedRemovalVersion?: string;
  }>;
// ...
}